Added multiple nodes and refactored

+++ src/test/java/org/kd/functional_interface/ChurchNumeralTest.java
@@ -3,7 +3,8 @@ package org.kd.functional_interface;
 import org.junit.jupiter.api.Test;
 import org.kd.functional_interface.custom.ChurchNumeral;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 
 public class ChurchNumeralTest {
 
@@ -16,7 +17,7 @@ public class ChurchNumeralTest {
     public void testApply2() {
         var five = ChurchNumeral.valueOf(5);
         System.out.println(five.apply(Math::sqrt, Double.MAX_VALUE));
-        assertEquals((Double) 4.2949672959999995E9, (Double) five.apply(Math::sqrt, Double.MAX_VALUE));
+        assertEquals((Double) 4.2949672959999995E9, five.apply(Math::sqrt, Double.MAX_VALUE));
     }
 
     public static void main(String[] a) {
